refactor(iocuak-common): move Newable to common models
