+++ packages/iocuak-common/src/common/models/Newable.ts
