+++ packages/iocuak-common/src/service/models/ServiceId.ts
@@ -1,3 +1,3 @@
-import { Newable } from './Newable';
+import { Newable } from '../../common/models/Newable';
 
 export type ServiceId = string | symbol | Newable;
